Apply functions to elements of a cycle

Function capply() means “cycle apply” and is modelled on lapply(). It applies a function to every element in the cycles of its argument.

Usage

capply(X, fun, ...)

Arguments

X

Permutation, coerced to cycle

fun

Function to be applied to each element of X

...

Further arguments to fun()

Details

This function is just a convenience wrapper really.

Value

Returns a permutation in cycle form

Author

Robin K. S. Hankin

Note

Function allcyc() is a nice application of capply().

See also

Examples

(x <- rperm())
#>  [1] (17)(254)   (1724)(356) (1476532)   (16)(45)    (1562)(34)  (267453)   
#>  [7] (264)(357)  (143)(2657) (23754)     (15)(367)  
#> [coerced from word form]
capply(x,function(x){x+100})
#>  [1] (101,107)(102,105,104)         (101,107,102,104)(103,105,106)
#>  [3] (101,104,107,106,105,103,102)  (101,106)(104,105)            
#>  [5] (101,105,106,102)(103,104)     (102,106,107,104,105,103)     
#>  [7] (102,106,104)(103,105,107)     (101,104,103)(102,106,105,107)
#>  [9] (102,103,107,105,104)          (101,105)(103,106,107)        
capply(x,rev)
#>  [1] (17)(245)   (1427)(365) (1235674)   (16)(45)    (1265)(34)  (235476)   
#>  [7] (246)(375)  (134)(2756) (24573)     (15)(376)  

all(is.id(capply(x,rev)*x))  # should be TRUE
#> [1] TRUE

capply(rcyc(20,5,9),sort)
#>  [1] (12389) (24579) (23469) (12489) (45679) (36789) (12468) (23579) (12579)
#> [10] (12789) (45678) (12679) (13456) (13458) (34789) (36789) (34679) (12346)
#> [19] (12389) (12367)

capply(rcyc(20,5,9),sample)  # still 5-cycles
#>  [1] (14326) (15246) (27359) (29453) (18762) (14673) (13684) (35769) (26948)
#> [10] (13768) (12693) (35879) (18239) (17285) (29637) (17234) (38975) (14985)
#> [19] (19635) (19756)

capply(cyc_len(1:9),\(x)x[x>4])
#> [1] ()      ()      ()      ()      ()      (56)    (567)   (5678)  (56789)

capply(cyc_len(1:9),\(x)x^2)
#> [1] ()                        (1,4)                    
#> [3] (1,4,9)                   (1,4,9,16)               
#> [5] (1,4,9,16,25)             (1,4,9,16,25,36)         
#> [7] (1,4,9,16,25,36,49)       (1,4,9,16,25,36,49,64)   
#> [9] (1,4,9,16,25,36,49,64,81)