Random permutations

Function rperm() creates a word object of random permutations. Function rcyc() creates random permutations comprising a single (group-theoretic) cycle of a specified length. Functions r1cyc() and rgs1() are low-level helper functions.

Usage

rperm(n=10,r=7,moved=NA)
rcyc(n,len,r=len)
r1cyc(len,vec)
rgs1(s)
rgivenshape(n,s,size=sum(s))

Arguments

n

Number of permutations to create

r

Size of permutations

len

Length of cycles in rcyc() and r1cyc()

moved

In function rperm(), integer specifying how many elements can move (that is, how many elements do not map to themselves), with default NA meaning to choose a permutation at random. This is useful if you want a permutation that has a compact cycle representation

vec

Vector of integers to generate a cycle from

s,size

Shape and size of permutations to create

Value

Returns an object of class word

Author

Robin K. S. Hankin

Note

Argument moved specifies a maximum number of elements that do not map to themselves; the actual number of non-fixed elements might be lower (as some elements might map to themselves). You can control the number of non-fixed elements precisely with argument len of function rcyc(), although this will give only permutations with a single (group-theoretic) cycle.

Argument s of function rgivensize() can include 1s (ones). Although length-one cycles are dropped from the resulting permutation, it is sometimes useful to include them to increase the size of the result, see examples.

In function rgivenshape(), if primary argument n is a vector of length greater than 1, it is interpreted as the shape of the permutation, and a single random permutation is returned.

See also

Examples

rperm()
#>  [1] (137624)    (26)(354)   (17)(25436) (1364)(275) (14)(25376) (1375)(46) 
#>  [7] (154637)    (264375)    (137)(2546) (1274)     
#> [coerced from word form]
as.cycle(rperm(30,9))
#>  [1] (13296)        (158742693)    (198)(23745)   (238)(57)      (36945)(78)   
#>  [6] (24539)(78)    (1597284)(36)  (18642)(37)    (129635784)    (147)(29835)  
#> [11] (1852)(34796)  (12)(376495)   (187)(24396)   (14386)(579)   (1258)(3749)  
#> [16] (25864937)     (173624)(58)   (134789)(256)  (1597684)(23)  (1638472)     
#> [21] (276583)       (185)(246)(39) (197365)(28)   (14276839)     (1639257)(48) 
#> [26] (175623)(49)   (137962548)    (278)(39)      (153)(274)(68) (16842735)    
rperm(10,9,2)
#>  [1] (79) (37) (68) (13) (38) (19) ()   (79) (19) (37)
#> [coerced from word form]

rcyc(20,5)
#>  [1] (12543) (12534) (14352) (15432) (14325) (12543) (14235) (14253) (12345)
#> [10] (15234) (13254) (15243) (15243) (13524) (14532) (14253) (13542) (12345)
#> [19] (15324) (13254)
rcyc(20,5,9)
#>  [1] (15269) (17654) (45689) (19267) (13689) (35796) (18495) (18527) (15243)
#> [10] (38564) (19753) (19532) (19732) (12487) (27359) (36978) (15937) (15378)
#> [19] (25486) (28569)

rgivenshape(10,c(2,3))   # size 5
#>  [1] (145)(23) (145)(23) (13)(245) (135)(24) (132)(45) (124)(35) (124)(35)
#>  [8] (13)(254) (134)(25) (14)(235)
rgivenshape(10,c(2,3,1,1)) # size 7
#>  [1] (172)(36) (164)(25) (175)(36) (134)(26) (257)(36) (263)(45) (12)(357)
#>  [8] (265)(34) (234)(57) (27)(465)

rgivenshape(1:9)
#> [1] (1,34,4,44,27,16,14,12)(2,19,17,3,36,45,38,37,35)(5,6,42,24,8,28)(7,40,25,9,41)(10,23,22,15,39,33,32)(11,20,18)(13,31)(21,26,30,43)

allpermslike(rgivenshape(c(1,1,3,4)))
#>  [1] (178)(3469) (187)(3469) (178)(3496) (187)(3496) (178)(3694) (187)(3694)
#>  [7] (178)(3649) (187)(3649) (178)(3946) (187)(3946) (178)(3964) (187)(3964)